fix: add hamburger button as part of the header
Include the hamburger button as an element of the header (possible form of refactoring)

The navbar does not fill in the entirety of the map's width; directly setting the width of the navbar greatly reduces the hamburger button's toggle response area
