+++ ConcordiaMaps/components/NavBar.js
@@ -1,23 +1,12 @@
 import React, { useState } from "react";
-import { View, TouchableOpacity, Text, Animated, Alert } from "react-native";
+import { View, TouchableOpacity, Text, Alert } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import styles from "../styles";
 import ShuttleSchedule from "./ShuttleSchedule";
 
 function NavBar() {
-  const [isOpen, setIsOpen] = useState(false);
   const [isScheduleVisible, setIsScheduleVisible] = useState(false);
-  const animation = useState(new Animated.Value(0))[0];
   const navigation = useNavigation();
-
-  const toggleMenu = () => {
-    const toValue = isOpen ? 0 : 1;
-    Animated.timing(animation, {
-      toValue,
-      duration: 300,
-      useNativeDriver: true,
-    }).start(() => setIsOpen(!isOpen));
-  };
 
   const handlePress = (item) => {
     if (item === "Get directions") {
@@ -27,47 +16,30 @@ function NavBar() {
     }
   };
 
-  const translateX = animation.interpolate({
-    inputRange: [0, 1],
-    outputRange: [-270, 0],
-  });
-
   return (
     <View style={styles.navbar}>
-      <TouchableOpacity
-        onPress={toggleMenu}
-        style={styles.hamburger}
-        testID="hamburger-button"
-      >
-        <View style={styles.hamburgerLine}></View>
-        <View style={styles.hamburgerLine}></View>
-        <View style={styles.hamburgerLine}></View>
+      <TouchableOpacity onPress={() => handlePress("Login")}>
+        <Text style={styles.menuItem}>Login</Text>
       </TouchableOpacity>
 
-      <Animated.View style={[styles.menu, { transform: [{ translateX }] }]}>
-        <TouchableOpacity onPress={() => handlePress("Login")}>
-          <Text style={styles.menuItem}>Login</Text>
-        </TouchableOpacity>
+      <TouchableOpacity onPress={() => handlePress("Get directions")}>
+        <Text style={styles.menuItem}>Get directions</Text>
+      </TouchableOpacity>
 
-        <TouchableOpacity onPress={() => handlePress("Get directions")}>
-          <Text style={styles.menuItem}>Get directions</Text>
-        </TouchableOpacity>
+      <TouchableOpacity onPress={() => handlePress("Outdoor Points of Interest")}>
+        <Text style={styles.menuItem}>Outdoor Points of Interest</Text>
+      </TouchableOpacity>
 
-        <TouchableOpacity
-          onPress={() => handlePress("Outdoor Points of Interest")}>
-          <Text style={styles.menuItem}>Outdoor Points of Interest</Text>
-        </TouchableOpacity>
-        
-        <TouchableOpacity onPress={() => handlePress("Smart Planner")}>
-          <Text style={styles.menuItem}>Smart Planner</Text>
-        </TouchableOpacity>
+      <TouchableOpacity onPress={() => handlePress("Smart Planner")}>
+        <Text style={styles.menuItem}>Smart Planner</Text>
+      </TouchableOpacity>
 
-        <TouchableOpacity
-          onPress={() => setIsScheduleVisible(true)}
-          testID="shuttle-schedule-modal">
-          <Text style={styles.menuItem}>Shuttle Schedule</Text>
-        </TouchableOpacity>
-      </Animated.View>
+      <TouchableOpacity
+        onPress={() => setIsScheduleVisible(true)}
+        testID="shuttle-schedule-modal"
+      >
+        <Text style={styles.menuItem}>Shuttle Schedule</Text>
+      </TouchableOpacity>
 
       {/* Shuttle Schedule Popup */}
       <ShuttleSchedule
